move external session migration into crate
