+++ codex-rs/external-agent-sessions/Cargo.toml
@@ -16,7 +16,9 @@ workspace = true
 chrono = { workspace = true }
 codex-protocol = { workspace = true }
 codex-utils-output-truncation = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
+sha2 = { workspace = true }
 
 [dev-dependencies]
 codex-app-server-protocol = { workspace = true }
